fix: updates how the context name is chosen when adding new contexts.
If a context name is provided via the --context-name argument, all contexts added will use that name (myContext, myContext-1, etc.). If no context name is provided as an argument, then the context name from the config file provided will be used. If no context name is found in the config file, then a new name will be generated based on the cluster and user name used by the new context (i.e. myCluster-myUser).

Refs: #43